grub2: Save default entry as index instead of title
Sometimes the new installed system could not change the default entry
for that the entry title in /boot/grub/grub.cfg it out of date in some
reason. For example, if user install system under bootable usb disk, the
main disk number may change to /dev/sdb but if you boot to the system it
will revert to /dev/sda1, so the os-probe entry titles which contains
such information will be out of date.

Now we save the default entry as index instead of title to avoid such
issue.

Resolves: https://bugzilla.deepin.io/show_bug.cgi?id=7065

Change-Id: I0bfb60f806aec0d86573d0d848b5e17ff2138331