Add remaining time to battery indicator
Add a QDBusConnection to org.freedesktop.UPower to DBusPower in
dbuspower.cpp.

Add a QDBusInterface (m_uPowerInter) to PowerPlugin in powerplugin.h to
retrieve the correct battery device (calling 'EnumerateDevices'). Add a
QDBusInterface (m_uBatteryDeviceInter) to PowerPlugin in powerplugin.h
to gather battery time to empty and to fully recharged, and display it
in the power widget tip.

Add BatteryState enum in powerplugin.h to replace BATTERY_DISCHARED and
BATTERY_FULL macros, and add support for all the battery states provided
by UPower.

Change-Id: Ia4677298a1d13cab1e1a03035dbe4a2e660f994b