feat: Initial packit setup
This commit contains the specfile for building the official package for Fedora
with a Packit setup.

Ultimately, a unified specfile is targeted for Fedora and any other rpm-based
distributions, e.g. openEuler.

And Packit(https://packit.dev/) is a tool for maintaining specfile within
upstream source. It requires a simple config file(.packit.yaml).

Log:
Change-Id: I976a21b9e63596b5b73d21f3edea3491053db192
Reviewed-on: http://gerrit.uniontech.com/c/dde-daemon/+/8249
Reviewed-by: <mailman@uniontech.com>
Reviewed-by: songwentai <songwentai@uniontech.com>
Tested-by: <mailman@uniontech.com>