fix: click undock menu which make fixed_plugin hide always
The removal will actually remove the plugin, which will make the removed plugin inaccessible,
making it impossible to set the plug-in to the dock in the dde-control-center.

Therefore, instead of calling the remove method, set the plugin to hide from the dock.

log: make fixed_plugin hide instead of removal when click ondock